[Refactor] Split basic test
* Moved some tests from basic test to `sync_and_flush_test`.
