+++ scripts/runtests.sh
@@ -24,6 +24,7 @@ $VALGRIND ./tests/table_test --abort-on-failure
 $VALGRIND ./tests/table_lookup_booster_test --abort-on-failure
 
 $VALGRIND ./tests/basic_op_test --abort-on-failure
+$VALGRIND ./tests/sync_and_flush_test --abort-on-failure
 $VALGRIND ./tests/nearest_search_test --abort-on-failure
 $VALGRIND ./tests/seq_itr_test --abort-on-failure
 $VALGRIND ./tests/key_itr_test --abort-on-failure
